Moved starknet_keccak and allowed_libfuncs to `starknet-classes` crate.
Moved validation to ContractClass to allow this.

commit-id:a4597890
