+++ crates/bin/starknet-compile/Cargo.toml
@@ -12,3 +12,4 @@ clap.workspace = true
 
 cairo-lang-compiler = { path = "../../cairo-lang-compiler", version = "2.5.3" }
 cairo-lang-starknet = { path = "../../cairo-lang-starknet", version = "2.5.3" }
+cairo-lang-starknet-classes = { path = "../../cairo-lang-starknet-classes", version = "2.5.3" }
