+++ crates/bin/starknet-sierra-compile/Cargo.toml
@@ -14,4 +14,5 @@ serde_json.workspace = true
 
 cairo-lang-sierra = { path = "../../cairo-lang-sierra", version = "2.5.3" }
 cairo-lang-starknet = { path = "../../cairo-lang-starknet", version = "2.5.3" }
+cairo-lang-starknet-classes = { path = "../../cairo-lang-starknet-classes", version = "2.5.3" }
 cairo-lang-utils = { path = "../../cairo-lang-utils", version = "2.5.3", features = ["serde"] }
