+++ crates/bin/starknet-sierra-compile/src/main.rs
@@ -1,9 +1,9 @@
 use std::fs;
 
 use anyhow::Context;
-use cairo_lang_starknet::allowed_libfuncs::{validate_compatible_sierra_version, ListSelector};
 use cairo_lang_starknet::casm_contract_class::CasmContractClass;
 use cairo_lang_starknet::contract_class::{ContractClass, ContractEntryPoints};
+use cairo_lang_starknet_classes::allowed_libfuncs::ListSelector;
 use cairo_lang_utils::bigint::BigUintAsHex;
 use clap::Parser;
 use serde::Deserialize;
@@ -62,7 +62,7 @@ fn main() -> anyhow::Result<()> {
         entry_points_by_type,
         abi: None,
     };
-    validate_compatible_sierra_version(&contract_class, list_selector)?;
+    contract_class.validate_version_compatible(list_selector)?;
     let casm_contract =
         CasmContractClass::from_contract_class(contract_class, args.add_pythonic_hints)
             .with_context(|| "Compilation failed.")?;
