+++ crates/bin/starknet-sierra-upgrade-validate/Cargo.toml
@@ -15,4 +15,5 @@ rayon.workspace = true
 indicatif = "0.17.7"
 
 cairo-lang-starknet = { path = "../../cairo-lang-starknet", version = "2.5.3" }
+cairo-lang-starknet-classes = { path = "../../cairo-lang-starknet-classes", version = "2.5.3" }
 cairo-lang-utils = { path = "../../cairo-lang-utils", version = "2.5.3", features = ["serde"] }
