+++ crates/cairo-lang-starknet-classes/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "cairo-lang-starknet-classes"
+version.workspace = true
+edition.workspace = true
+repository.workspace = true
+license-file.workspace = true
+description = "Starknet definitions for contract classes."
+
+[dependencies]
+cairo-lang-sierra = { path = "../cairo-lang-sierra", version = "2.5.3" }
+num-bigint = { workspace = true, default-features = true }
+serde = { workspace = true, default-features = true }
+serde_json.workspace = true
+sha3.workspace = true
+smol_str.workspace = true
+thiserror.workspace = true
+
+[dev-dependencies]
+env_logger.workspace = true
+pretty_assertions.workspace = true
+test-case.workspace = true
+test-log.workspace = true
