+++ crates/cairo-lang-starknet-classes/src/allowed_libfuncs_lists/all.json
