+++ crates/cairo-lang-starknet-classes/src/keccak_test.rs
@@ -1,0 +1,9 @@
+use crate::keccak::starknet_keccak;
+
+#[test]
+fn test_starknet_keccak() {
+    assert_eq!(
+        format!("0x{:x}", starknet_keccak("__execute__".as_bytes())),
+        "0x15d40a3d6ca2ac30f4031e42be28da9b056fef9bb7357ac5e85627ee876e5ad",
+    )
+}
