+++ crates/cairo-lang-starknet-classes/src/lib.rs
@@ -1,0 +1,9 @@
+//! Starknet classes definitions.
+//!
+//! Starknet is a smart contract platform that enables developers to build and deploy smart
+//! contracts on a permissionless Layer 2 network, secured by Ethereum using validity proofs.
+//!
+//! Learn more at [starkware.io](http://starknet.io/).
+
+pub mod allowed_libfuncs;
+pub mod keccak;
