+++ crates/cairo-lang-starknet/src/contract.rs
@@ -20,6 +20,7 @@ use cairo_lang_semantic::Expr;
 use cairo_lang_sierra::ids::FunctionId;
 use cairo_lang_sierra_generator::db::SierraGenGroup;
 use cairo_lang_sierra_generator::replace_ids::SierraIdReplacer;
+use cairo_lang_starknet_classes::keccak::starknet_keccak;
 use cairo_lang_syntax::node::helpers::{GetIdentifier, PathSegmentEx, QueryAttrs};
 use cairo_lang_syntax::node::TypedSyntaxNode;
 use cairo_lang_utils::extract_matches;
@@ -27,9 +28,7 @@ use cairo_lang_utils::ordered_hash_map::{
     deserialize_ordered_hashmap_vec, serialize_ordered_hashmap_vec, OrderedHashMap,
 };
 use itertools::chain;
-use num_bigint::BigUint;
 use serde::{Deserialize, Serialize};
-use sha3::{Digest, Keccak256};
 use {cairo_lang_lowering as lowering, cairo_lang_semantic as semantic};
 
 use crate::aliased::Aliased;
@@ -51,17 +50,6 @@ impl ContractDeclaration {
     pub fn module_id(&self) -> ModuleId {
         ModuleId::Submodule(self.submodule_id)
     }
-}
-
-/// A variant of eth-keccak that computes a value that fits in a Starknet field element.
-pub fn starknet_keccak(data: &[u8]) -> BigUint {
-    let mut hasher = Keccak256::new();
-    hasher.update(data);
-    let mut result = hasher.finalize();
-
-    // Truncate result to 250 bits.
-    *result.first_mut().unwrap() &= 3;
-    BigUint::from_bytes_be(&result)
 }
 
 /// Returns the contract declaration of a given module if it is a contract module.
