+++ crates/cairo-lang-starknet/src/felt252_serde.rs
@@ -19,6 +19,7 @@ use cairo_lang_sierra::program::{
     Function, FunctionSignature, GenericArg, Invocation, LibfuncDeclaration, Param, Program,
     Statement, StatementIdx, TypeDeclaration,
 };
+use cairo_lang_starknet_classes::keccak::starknet_keccak;
 use cairo_lang_utils::bigint::BigUintAsHex;
 use cairo_lang_utils::ordered_hash_set::OrderedHashSet;
 use cairo_lang_utils::unordered_hash_map::UnorderedHashMap;
@@ -29,7 +30,6 @@ use smol_str::SmolStr;
 use thiserror::Error;
 
 use crate::compiler_version::VersionId;
-use crate::contract::starknet_keccak;
 use crate::felt252_vec_compression::{compress, decompress};
 
 #[cfg(test)]
