+++ crates/cairo-lang-starknet/src/inline_macros/selector.rs
@@ -2,10 +2,9 @@ use cairo_lang_defs::plugin::{
     InlineMacroExprPlugin, InlinePluginResult, NamedPlugin, PluginDiagnostic, PluginGeneratedFile,
 };
 use cairo_lang_semantic::extract_macro_single_unnamed_arg;
+use cairo_lang_starknet_classes::keccak::starknet_keccak;
 use cairo_lang_syntax::node::db::SyntaxGroup;
 use cairo_lang_syntax::node::{ast, TypedSyntaxNode};
-
-use crate::contract::starknet_keccak;
 
 /// Macro for expanding a selector to a string literal.
 #[derive(Debug, Default)]
